+++ bot/commands/info/commands.js
@@ -1,3 +1,5 @@
 module.exports = function(bot, db) {
-    bot.sendChat("Check out all of the commands at my repo: https://github.com/franciscog/DerpyBot#commands");
+  bot.sendChat("Check out all of the commands here: https://github.com/franciscog/DerpyBot#commands");
 };
+
+module.exports.extraCommands = ['help'];