+++ bot/commands/info/help.js
@@ -1,3 +1,0 @@
-module.exports = function(bot, db) {
-    bot.sendChat("Here are a few commands to get you started: !rules, !dubx, !eta, !link, !tune, !love");
-};